Update entry transformation implementation
This gets rid of the groupBy copy/paste function, transforms the entries
in one step to avoid going through the data multiple times. It also uses
the moment date objects in the data structure to avoid parsing and
converting back and forth between strings and moment date objects when
comparing and sorting.